+++ packages/expo-router/CHANGELOG.md
@@ -11,6 +11,7 @@
 ### 💡 Others
 
 - bump e2e tests to React 19 and fix lint ([#36344](https://github.com/expo/expo/pull/36344) by [@EvanBacon](https://github.com/EvanBacon))
+- Refactor route state to utilize useStateForPath() ([#36199](https://github.com/expo/expo/pull/36199) by [@marklawlor](https://github.com/marklawlor))
 
 ## 5.0.2-preview.5 — 2025-04-22